% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_summits.R
\name{annotate_summits}
\alias{annotate_summits}
\title{annotate_summits}
\usage{
annotate_summits(
  .peak_points,
  .peak_sizes,
  .min_kmerpair_cov,
  .max_kmerpair_cov,
  col = "red"
)
}
\description{
\code{annotate_summits}
}
