% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.nrdW.R
\name{bw.nrdW}
\alias{bw.nrdW}
\title{bw.nrdW}
\usage{
bw.nrdW(x, w)
}
\description{
\code{bw.nrdW} is a function that mimics bw.nrd0 - Silverman's ‘rule of thumb’ for estimating the kernel width, but using weighted values;
this can be useful if one wants to mimics kerned smoothing to a histogram using counts as weights
}
