% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_histogram.R
\name{coverage_histogram}
\alias{coverage_histogram}
\title{coverage_histogram}
\usage{
coverage_histogram(.cov_tab, bins, xlim, column, horiz, .col)
}
\description{
\code{coverage_histogram} plots a histogram from a table that contains both values (what is normally x) and their respective frequencies (a column called 'freq');
}
