% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_annotation_grid_fit.R
\name{create_smudge_container}
\alias{create_smudge_container}
\title{grid_fit}
\usage{
create_smudge_container(cov, cov_tab, smudge_filtering_threshold)
}
\description{
\code{create_smudge_container} creates a smudge container (a list if points and their respective)
\code{run_replicate} creates
\code{get_centrality} calculates distance of the closest edge to the center
\code{test_grid_of_coverages} given minumum and maximum coverage, performs a grid search
}
