% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_1n_coverage_1d_subsets.R
\name{estimate_1n_coverage_1d_subsets}
\alias{estimate_1n_coverage_1d_subsets}
\title{estimate_1n_coverage_1d_subsets}
\usage{
estimate_1n_coverage_1d_subsets(.cov_tab)
}
\description{
\code{estimate_1n_coverage_1d_subsets} is used as the first proxy of 1n coverage. The estimate is subsequentially refined by the brightest smudge.
}
