% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_1n_coverage_highest_peak.R
\name{estimate_1n_coverage_highest_peak}
\alias{estimate_1n_coverage_highest_peak}
\title{estimate_1n_coverage_highest_peak}
\usage{
estimate_1n_coverage_highest_peak(.filt_peak_sizes, .cov_tab, .draft_1n)
}
\description{
\code{estimate_1n_coverage_highest_peak}
}
