% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_1d_peaks.R
\name{get_1d_peaks}
\alias{get_1d_peaks}
\title{get_1d_peaks}
\usage{
get_1d_peaks(
  .cov_tab,
  .num_of_peaks = 3,
  .adjust = 10,
  .peak_frame = data.frame(),
  .depth = 1
)
}
\description{
\code{get_1d_peaks} is a legacy code
}
