% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_ramp.R
\name{get_col_ramp}
\alias{get_col_ramp}
\title{get_col_ramp}
\usage{
get_col_ramp(.args, delay = 0)
}
\description{
\code{get_col_ramp} returns a colour ramp used by smudgeplots; 
args need to contain args$col_ramp and args$invert_cols features
the first needs to be a ramp-generating function (e.g. "viridis", "grey.colors", "magma" or "mako")
args$invert_cols is just TRUE or FALSE (if to revert the colours in the palete)
}
