% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smudge_container.R
\name{get_smudge_container}
\alias{get_smudge_container}
\title{get_smudge_container}
\usage{
get_smudge_container(.cov_tab, .nbins = 20, .xlim = c(0, 0.5), .ylim = NA)
}
\description{
\code{get_smudge_container} is inspired by https://stackoverflow.com/a/18103689/2962344
}
