% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_genome_structure.R
\name{guess_genome_structure}
\alias{guess_genome_structure}
\title{guess_genome_structure}
\usage{
guess_genome_structure(.filt_peak_sizes_line, .n)
}
\description{
\code{guess_genome_structure}
}
