% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_agregation.R
\name{peak_agregation}
\alias{peak_agregation}
\title{peak_agregation}
\usage{
peak_agregation(.smudge_container, .L = NA)
}
\arguments{
\item{.smudge_container}{- a smudge_container object; a list with `x`,`y` axies of the 2d histogram and `dens` matrix of 2h histogram values.
Usually output of function \code{get_smudge_container}}
}
\description{
\code{peak_agregation} is the function that agregates tiles from the smudge_container
into smudges - individual distributions
}
