% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alt.R
\name{plot_alt}
\alias{plot_alt}
\title{plot_alt}
\usage{
plot_alt(cov_tab, ylim, colour_ramp, log = F)
}
\description{
\code{plot_alt} is an alternative plotting introduced for 0.3.0 oriel. Instead of summing all the coverages falling within boundaries of squares, we make a tiling that will cover each individual point
}
