% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expected_haplotype_structure.R
\name{plot_expected_haplotype_structure}
\alias{plot_expected_haplotype_structure}
\title{plot_expected_haplotype_structure}
\usage{
plot_expected_haplotype_structure(
  .n,
  .peak_sizes,
  .adjust = F,
  .cex = 1.3,
  xmax = 0.49
)
}
\description{
\code{plot_expected_haplotype_structure} adds genome scture labels to a smudgeplot
}
