% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\name{plot_histograms}
\alias{plot_histograms}
\title{plot_histograms}
\usage{
plot_histograms(
  .cov_tab,
  .fig_title = NA,
  .cex = 1.4,
  .col = NA,
  .ylim = NA,
  .bins = 100
)
}
\description{
\code{plot_histograms} makes 2 plots - histograms of the both margins of smudge plot
}
