% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seq_error_line.R
\name{plot_seq_error_line}
\alias{plot_seq_error_line}
\title{plot_seq_error_line}
\usage{
plot_seq_error_line(.cov_tab, .L = NA, .col = "black")
}
\description{
\code{plot_seq_error_line} add error line to plot it's L / cov
}
