% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smudgeplot.R
\name{plot_smudgeplot}
\alias{plot_smudgeplot}
\title{plot_smudgeplot}
\usage{
plot_smudgeplot(.k, .n, .colour_ramp, .cex = 1.4)
}
\arguments{
\item{.k}{- the matrix of densities or list of x,y axies and z, the matrix. Usually output of function}

\item{.n}{- the 1n coverage, the coverage of the haploid genome (can be estimated by TODO function)}

\item{.colour_ramp}{- the colour pallete for the heat colours}

\item{.cex}{- the size of the axis labels [default 1.4]}
}
\description{
\code{plot_smudgeplot} is the core smudgeplot function to plot the core of the smudgeplot
the cental 2D histogram/ landscape plot / heatmap / howeveryouwantyoucallthis plot
}
\author{
Kamil Jaron \email{kamiljaron at gmail.com}
}
