% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_structure_representation.R
\name{reduce_structure_representation}
\alias{reduce_structure_representation}
\title{reduce_structure_representation}
\usage{
reduce_structure_representation(.peak_sizes)
}
\arguments{
\item{.peak_sizes}{- a table with peak sizes, the truly important column in there is just the "structure" one}
}
\description{
\code{reduce_structure_representation} is the function that collapses labels loger than 4 characters into 4 character representations
}
