% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_summary.R
\name{smudge_summary}
\alias{smudge_summary}
\title{smudge_summary}
\usage{
smudge_summary(.filename, ...)
}
\description{
\code{smudge_summary} a function that is used to report all the summary output messages
like this I will be able to easily redirect the waring message to file as well
}
