% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smudge_warn.R
\name{smudge_warn}
\alias{smudge_warn}
\title{smudge_warn}
\usage{
smudge_warn(.filename, ...)
}
\description{
\code{smudge_warn} a function that is used to report all the wanrning messages
like this I will be able to easily redirect the waring message to file as well
}
