% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_minor_variant_cov.R
\name{transform_minor_variant_cov}
\alias{transform_minor_variant_cov}
\title{transform_minor_variant_cov}
\usage{
transform_minor_variant_cov(.x, .smudge_container)
}
\description{
\code{transform_minor_variant_cov}
to transform coordinates of smudgeplot to minor_variant_rel_cov
}
