% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pair_cov.R
\name{transform_pair_cov}
\alias{transform_pair_cov}
\title{transform_pair_cov}
\usage{
transform_pair_cov(.y, .smudge_container)
}
\description{
\code{transform_pair_cov}
to transform coordinates of smudgeplot to total_pair_cov
}
