% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.quantile.R
\name{wtd.quantile}
\alias{wtd.quantile}
\title{wtd.quantile & wtd.iqr}
\usage{
wtd.quantile(x, q = 0.25, weight = NULL)
}
\description{
\code{wtd.quantile}
calculates a quantile using "wighted data", specifically designed for "histogram-like" data
\code{wtd.iqr}
0.75 - 0.25 weighted quantile
}
